TINY-9977: Separated the schema rule parser
