+++ modules/tinymce/src/core/main/ts/schema/CustomElementsRuleParser.ts
@@ -1,0 +1,25 @@
+import { Arr } from '@ephox/katamari';
+
+import * as SchemaUtils from './SchemaUtils';
+
+export interface CustomElementRules {
+  readonly inline: boolean;
+  readonly cloneName: 'div' | 'span';
+  readonly name: string;
+}
+
+export const parseCustomElementsRules = (value: string): CustomElementRules[] => {
+  const customElementRegExp = /^(~)?(.+)$/;
+  return Arr.bind(SchemaUtils.split(value, ','), (rule) => {
+    const matches = customElementRegExp.exec(rule);
+    if (matches) {
+      const inline = matches[1] === '~';
+      const cloneName = inline ? 'span' : 'div';
+      const name = matches[2];
+
+      return [{ inline, cloneName, name }];
+    } else {
+      return [];
+    }
+  });
+};
